+++ .gitignore
@@ -1,2 +1,3 @@
 .vscode/settings.json
 node_modules/
+coverage
