+++ src/get-package-info.js
